% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datey.R
\name{as.double.datey}
\alias{as.double.datey}
\alias{as.integer.datey}
\title{A \code{datey} as years}
\usage{
\method{as.double}{datey}(x, ...)

\method{as.integer}{datey}(x, ...)
}
\arguments{
\item{x}{The \code{datey} to convert to years.}

\item{...}{Further arguments to be passed from or to other methods.}
}
\description{
A \code{datey} converted to years.

For example,
\itemize{
\item the \emph{start} of 2000-01-01 (or, equivalently, the \emph{end} of 1999-12-31),
results in \code{2000}, and
\item the \emph{middle} of the calendar year 2000 results in \code{2000.5}.
}

A \code{datey} converted to integer years.

For example, 2000-01-01.0 and 2000-12-31.0 both result in \code{2000}.
}
