% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/durationy.R
\name{as.double.durationy}
\alias{as.double.durationy}
\alias{as.integer.durationy}
\title{A \code{durationy} as years}
\usage{
\method{as.double}{durationy}(x, ...)

\method{as.integer}{durationy}(x, ...)
}
\arguments{
\item{x}{The \code{durationy} to convert to years.}

\item{...}{Further arguments to be passed from or to other methods.}
}
\description{
A \code{durationy} converted to years.

Conversion to an integer rounds to the nearest whole year.
}
