% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datey.R
\name{as.ymdf}
\alias{as.ymdf}
\title{Deconstruct a \code{datey}}
\usage{
as.ymdf(datey)
}
\arguments{
\item{datey}{The \code{datey} to deconstruct.}
}
\description{
Deconstructs a \code{datey} into a list comprising \code{year}, \code{month}, \code{day} and
\code{day_fraction} components (in that order).
}
\details{
If \code{datey} is valid then the components will lie in the following intervals:
\itemize{
\item \code{year} an \code{integer} in [1000,3000),
\item \code{month} an \code{integer} in [1,12],
\item \code{day} an \code{integer} in [1,N], where N is the number of days
in the month specified by \code{year} and \code{month}, and
\item \code{day_fraction} a \code{double} in [0,1) representing the fraction of the day,
where, for instance, 0 means the start of the day and 0.5 means the middle of the day.
}
}
