% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datey.R
\name{as_datey}
\alias{as_datey}
\alias{as_datey.default}
\alias{as_datey.double}
\alias{as_datey.integer}
\alias{as_datey.character}
\alias{as_datey.Date}
\alias{as_datey.POSIXct}
\alias{as_datey.POSIXlt}
\alias{format.datey}
\alias{print.datey}
\title{Convert an object to a \code{datey}.}
\usage{
as_datey(x, ...)

\method{as_datey}{default}(x, ...)

\method{as_datey}{double}(x, ...)

\method{as_datey}{integer}(x, ...)

\method{as_datey}{character}(x, day_fraction = NULL, ...)

\method{as_datey}{Date}(x, day_fraction, ...)

\method{as_datey}{POSIXct}(x, ...)

\method{as_datey}{POSIXlt}(x, ...)

\method{format}{datey}(x, ...)

\method{print}{datey}(x, max = NULL, ...)
}
\arguments{
\item{x}{A vector.}

\item{...}{Other arguments.}

\item{day_fraction}{The fraction of the day, in [0,1].
0 means the start of the day,
0.5 means the middle of the day, and
1 means the end of the day
(which is identical to the start of the next day).}
}
\description{
This is an S3 generic. This package provides methods for the
following classes:
\itemize{
\item \code{double} and \code{integer}:
Values are interpreted as year, e.g. 2000.0. is the \emph{start of}
the year 2000.
\item \code{Date}: Note that .
\item \code{POSIXct}: Note that .
\item \code{POSIXlt}: Note that .
}
}
