% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/durationy.R
\name{as_durationy.character}
\alias{as_durationy.character}
\title{Parse text as a \code{durationy}}
\usage{
\method{as_durationy}{character}(x, strict = TRUE, blank_is_NA = FALSE, year_unit = "yr", ...)
}
\arguments{
\item{x}{Vector of text items to be parsed.}

\item{strict}{How non-compliant text (including values greater than
2000 in magnitude) should be handled.
\itemize{
\item If \code{strict} is \code{TRUE} then execution is stopped.
\item If \code{strict} is \code{FALSE} then \code{NA} is returned.
Defaults to \code{TRUE}.
}}

\item{blank_is_NA}{Whether blanks should be treated as \code{NA}.
Defaults to \code{FALSE}.}

\item{year_unit}{The year unit name to expect.
If not blank then the value is expected to be followed by a space and this unit text.
Cannot be more than 20 characters (UTF-8 bytes) or contain control characters.
Defaults to \code{"yr"}.}

\item{...}{Other arguments (not used in this package).}
}
\description{
If \code{day_fraction} \emph{is} provided then the text must be in ISO 8601 extended
format, i.e. "YYYY-MM-DD".

If \code{day_fraction} is \emph{not} provided then the text must be formatted as
"YYYY-MM-DD.FFF", where ".FFF" is the optional day fraction. This means that
e.g. "2000-01-01" represents the \emph{start} of 1 January 2000.

If \code{blank_is_NA} is \code{TRUE} then blanks are treated as \code{NA}.

If \code{strict} is \code{TRUE} (which is the default) then non-compliant text will
stop execution.

The lengths of vector arguments must be multiples of each other.
}
