% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/durationy.R
\name{as_durationy}
\alias{as_durationy}
\alias{as_durationy.default}
\alias{as_durationy.durationy}
\alias{as_durationy.integer}
\alias{as_durationy.double}
\title{Convert an object to a \code{durationy}}
\usage{
as_durationy(x, strict = TRUE, ...)

\method{as_durationy}{default}(x, strict = TRUE, ...)

\method{as_durationy}{durationy}(x, strict = TRUE, ...)

\method{as_durationy}{integer}(x, strict = TRUE, ...)

\method{as_durationy}{double}(x, strict = TRUE, ...)
}
\arguments{
\item{x}{A vector of the S3 class.}

\item{strict}{How years greater than 2000 in magnitude should be
handled.
\itemize{
\item If \code{strict} is \code{TRUE} -- the default -- then execution is stopped.
\item If \code{strict} is \code{FALSE} then \code{NA} is returned.
}

(NAs will result in NA regardless of this switch.)}

\item{...}{Other arguments (not used in this package).}
}
\description{
This is an S3 generic. This package provides methods for the
following classes:
\itemize{
\item \code{integer} -- the value is interpreted as the specified
number of years.
\item \code{double} -- the value is interpreted as the specified
number of years, rounded to fixed precision of a \code{durationy}. This means that
\code{as_durationy(0.5)} is precise but \code{as_durationy(0.01)} is not.
\item \code{durationy} -- value is unchanged.
}
}
