% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datey.R
\name{as_xxx_day}
\alias{as_xxx_day}
\alias{as_start_day}
\alias{as_mid_day}
\alias{as_end_day}
\title{Coerce a calendar year (including fractional part) or
another date type to a \code{datey} for the start, middle of end of the day}
\usage{
as_start_day(x, strict = TRUE)

as_mid_day(x, strict = TRUE)

as_end_day(x, strict = TRUE)
}
\arguments{
\item{x}{The argument to convert to a \code{datey}.}

\item{strict}{How invalid results (given non-NA inputs) should be handled.
\itemize{
\item If \code{strict} is \code{TRUE} -- the default -- then execution is stopped.
\item If \code{strict} is \code{FALSE} then \code{NA} is returned.
}

(NA inputs will result in NA regardless of this switch.)}
}
\description{
Accepted types are:
\itemize{
\item Numeric, interpreted as calendar year, with the fractional part
representing the fraction of the year. For instance, \code{datey(2000.5)} means
halfway though the year 2000.
\item The base R date types (\code{Date} and \code{POSIXct} and \code{POSIXlt}).
\item \code{character}, in ISO 8601 extended format, i.e. YYYY-MM-DD.
\item \code{datey}, which is interpreted as is but with the start, middle or end day
override.
}

Beware that \code{as_end_day()} will add a day to a \code{datey} that is already on a
day boundary, \emph{even if it was originally defined as an end day}.
}
\seealso{
Use \code{\link[=start_day]{start_day()}}, \code{\link[=mid_day]{mid_day()}} and \code{\link[=end_day]{end_day()}} to create a \code{datey}
direct from year, month and day.
}
