% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datey.R
\name{as_years_datey}
\alias{as_years_datey}
\alias{as.double.datey}
\alias{as.integer.datey}
\title{Convert a \code{datey} to calendar years (including fractional part)}
\usage{
\method{as.double}{datey}(x, ...)

\method{as.integer}{datey}(x, ...)
}
\arguments{
\item{x}{The \code{datey} to convert to years.}

\item{...}{Not used.}
}
\value{
A vector of \code{double}.
}
\description{
Converts a \code{datey} to calendar years,
including a fractional part that represents the
proportion of the calendar year that has elapsed.

For example, the middle of 2000-10-01 is precisely three-quarters through
the (leap) year 2000 and so \code{as.double(mid_day(2000,10,1))} results in
\code{2000.75}.

\code{as.numeric()} is the same as \code{as.double()}.

\code{as.integer()} gives the calendar year as an \code{integer}, e.g.
\code{as.integer(datey(2000.75))} is \code{2000}.
It is also the case that if \code{x} is a \code{datey} then
\code{as.integer(x)} is the same as \code{as.integer(as.double(x))}.
}
\examples{
t <- datey(2000.75)
t                   # 2000-10-01.5
as.double(t)        # 2000.75
as.numeric(t)       # 2000.75
as.integer(t)       # 2000
identical(as.integer(t), 2000L) # TRUE
}
\seealso{
\link{datey}, \link{as_years_durationy}, \link{ops}
}
