% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datey.R
\name{c.datey}
\alias{c.datey}
\title{Concatenate \code{datey} vectors}
\usage{
\method{c}{datey}(..., recursive = FALSE)
}
\arguments{
\item{...}{The items to combine}

\item{recursive}{Unused.}
}
\value{
\code{\link[=c]{c()}} returns a vector of \code{datey}s.

[cbind()] and [rbind()] return a matrix, data.frame or list with dimensions
}
\description{
Combines (flattens) \code{datey} vectors.

If the first element in \code{c(...)} is not a \code{datey} then this method will not
be called. For instance, \code{c(NA, as_datey("2000-01-01.0"))}
}
\note{
R currently only dispatches generic \code{c} to method \code{c.datey} if the
first argument is a \code{datey}.
}
\examples{
  c(as_datey(2000), as_datey("2020-01-01.0"))
  #cbind(1:6, as.datey(2001:2020))
  #rbind(1:6, as.datey(2001:2020))
}
\keyword{classes}
\keyword{manip}
