% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/durationy.R
\name{c.durationy}
\alias{c.durationy}
\title{Concatenate \code{durationy} vectors}
\usage{
\method{c}{durationy}(..., recursive = FALSE)
}
\arguments{
\item{...}{The items to combine}

\item{recursive}{Unused.}
}
\value{
\code{\link[=c]{c()}} returns a vector of \code{durationy}s.

[cbind()] and [rbind()] return a matrix, data.frame or list with dimensions
}
\description{
Combines (flattens) \code{durationy} vectors.

If the first element in \code{c(...)} is not a \code{durationy} then this method will not
be called. For instance, \code{c(NA, as_durationy(1))} is \emph{not} a \code{durationy}.
}
\note{
R currently only dispatches generic \code{c} to method \code{c.durationy} if the
first argument is a \code{durationy}.
}
\examples{
  c(as_durationy(1), as_durationy("2.5 yr"))
}
\keyword{classes}
\keyword{manip}
