% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datey.R
\docType{data}
\name{datey.NA}
\alias{datey.NA}
\alias{NA_datey_}
\alias{is.na.datey}
\alias{anyNA.datey}
\alias{valid_years_start}
\alias{valid_years_end}
\title{Handling invalid dates}
\format{
An object of class \code{datey} of length 1.

An object of class \code{integer} of length 1.

An object of class \code{integer} of length 1.
}
\usage{
NA_datey_

\method{is.na}{datey}(x)

\method{anyNA}{datey}(x, recursive = FALSE)

valid_years_start

valid_years_end
}
\arguments{
\item{x}{The \code{datey} to test for validity.}

\item{recursive}{Unused.}
}
\value{
A single logical value, \code{TRUE} or \code{FALSE}, never \code{NA} and never
anything other than a single value.
}
\description{
Valid dates within the datey system have calendar years in the interval
[1000,3000).

Dates outside this interval are treated as NA.

Use
\itemize{
\item \code{is.na()} to test whether \code{datey} is NA by element, and
\item \code{anyNA()} to test whether any element of a \code{datey} is NA.
}

The \code{datey} version of NA is \code{NA_datey_}.

For convenience, the following constants are also available:
\itemize{
\item \code{valid_years_start}: The first \emph{valid} calendar year (1000).
\item \code{valid_years_end}: The first \emph{invalid} calendar year after
\code{valid_years_start} (3000).
}

For performance reasons, intermediate calculations may not check for NAs.
}
\examples{
  x <- c(NA_datey_, as_datey(2000))
  is.na(x) # c(TRUE, FALSE)
  anyNA(x) # TRUE
}
\seealso{
as_datey
}
\keyword{NA}
\keyword{datasets}
