% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datey.R
\name{datey}
\alias{datey}
\alias{start_day}
\alias{mid_day}
\alias{end_day}
\title{Create a \code{datey}}
\usage{
datey(year, month, day, day_fraction)

start_day(year, month, day)

mid_day(year, month, day)

end_day(year, month, day)
}
\arguments{
\item{year}{Calendar year.
Valid years are from 1000 (inclusive) to 3000 (exclusive).
Doubles are truncated to integers.}

\item{month}{Month number in calendar year, with 1 representing January.
Doubles are truncated to integers.}

\item{day}{Day number in month, with 1 representing the first day of the month.
Doubles are truncated to integers.}

\item{day_fraction}{The fraction of the day, in [0,1].
0 means the start of the day,
0.5 means the middle of the day, and
1 means the end of the day
(which is identical to the start of the next day).}
}
\description{
Create a \code{datey} from
a year, month, day and, optionally, a day-fraction.

In general, prefer the explicit \code{\link[=start_day]{start_day()}}, \code{\link[=mid_day]{mid_day()}} and \code{\link[=end_day]{end_day()}}
versions.

To deconstruct a \code{datey}, use \code{\link[=as_ymdf]{as_ymdf()}}.
}
