% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datey.R
\name{datey_components}
\alias{datey_components}
\alias{to_ymdf}
\alias{$.datey}
\title{Get year, month, day or day_fraction breakdown of a \code{datey}}
\usage{
to_ymdf(x)

\method{$}{datey}(x, name)
}
\arguments{
\item{x}{The \code{datey} to be deconstructed.}

\item{name}{The name of the component for the list-like syntax.
Must be \code{year}, \code{month}, \code{day} or \code{day_fraction}.}
}
\value{
\code{to_ymdf()} returns a list of
integer vector \code{year},
integer vector \code{month},
integer vector \code{day}, and
double vector \code{day_fraction}, all with the same length.
The list-like syntax returns these components individually.
}
\description{
To extract the year, month, day or day_fraction breakdown of a \code{datey},
use either
\itemize{
\item the list-like syntax \verb{$year}, \verb{$month}, \verb{$day} or \verb{$day_fraction}
direct, or
\item if you need several components at once, \code{to_ymdf()}, which
returns an actual list of \code{year}, \code{month}, \code{day} and \code{day_fraction}.
}

In this breakdown,
\itemize{
\item \code{year} is an \code{integer} in [1000,3000],
\item \code{month} is an \code{integer} in [1,12],
\item \code{day} is an \code{integer} in [1,N], where N is the number of days
in the month specified by \code{year} and \code{month}, and
\item \code{day_fraction} is a \code{double} in [0,1) representing the fraction of the day,
where e.g. 0 means the start and 0.5 means the middle of the day.
}

If the \code{datey} was constructed using \code{end_day} or \code{day_fraction = 1} then
\code{to_ymdf()} will return the \emph{start} of the \emph{next} day with \code{day_fraction = 0}.
}
\examples{
  t <- datey(2001, 2, 3, 0.5)
  t
  to_ymdf(t)
  t$year
  t$month
  t$day
  t$day_fraction
}
\seealso{
\link{datey}, \link{text_from_datey}
}
