% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datey.R
\name{datey_properties}
\alias{datey_properties}
\alias{$.datey}
\title{Extract year, month, day or day_fraction from a \code{datey}}
\usage{
\method{$}{datey}(x, name)
}
\arguments{
\item{x}{The \code{datey}.}

\item{name}{Must be \code{year}, \code{month}, \code{day} or \code{day_fraction}.}
}
\value{
A vector of \code{integer} for \code{year}, \code{month} and \code{day};
a vector of \code{integer} for \code{day_fraction}.
}
\description{
Extract the year, month, day or day_fraction of a \code{datey} using the syntax
\verb{$year}, \verb{$month}, \verb{$day} or \verb{$day_fraction} respectively.

\verb{$year}, \verb{$month} and \verb{$day} are \code{integer}, and \verb{$day_fraction} is \code{double}.
}
\examples{
  t <- mid_day(2001, 2, 3)
  t$year
  t$month
  t$day
  t$day_fraction
}
\seealso{
If you need more than one component then \code{\link[=to_ymdf]{to_ymdf()}} may be more efficient.
}
