% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/durationy.R
\docType{data}
\name{durationy.NA}
\alias{durationy.NA}
\alias{NA_durationy_}
\alias{is.na.durationy}
\alias{anyNA.durationy}
\alias{valid_duration_years_max}
\title{Handling invalid durations}
\format{
An object of class \code{durationy} of length 1.

An object of class \code{integer} of length 1.
}
\usage{
NA_durationy_

\method{is.na}{durationy}(x)

\method{anyNA}{durationy}(x, recursive = FALSE)

valid_duration_years_max
}
\arguments{
\item{x}{The \code{durationy} to test for validity.}

\item{recursive}{Unused.}
}
\value{
A single logical value, \code{TRUE} or \code{FALSE}, never \code{NA} and never
anything other than a single value.
}
\description{
A duration within the datey system is valid if it represents 2000 years or
less.

Larger durations are treated as NA.

Use
\itemize{
\item \code{is.na()} to test whether \code{datey} is NA by element, and
\item \code{anyNA()} to test whether any element of a \code{datey} is NA.
}

The \code{durationy} version of NA is \code{NA_durationy_}.

For convenience, the constant \code{valid_max_duration} is the maximum valid
duration in years (2000).

For performance reasons, intermediate calculations may not check for NAs.
}
\examples{
  x <- c(NA_durationy_, as_durationy(1.5))
  is.na(x) # c(TRUE, FALSE)
  anyNA(x) # TRUE
}
\seealso{
as_durationy
}
\keyword{NA}
\keyword{datasets}
