% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/durationy.R
\name{format.durationy}
\alias{format.durationy}
\alias{print.durationy}
\title{Format or print a \code{durationy}}
\usage{
\method{format}{durationy}(x, include_plus = FALSE, use_true_minus = TRUE, year_unit = "yr", ...)

\method{print}{durationy}(
  x,
  include_plus = FALSE,
  use_true_minus = TRUE,
  year_unit = "yr",
  max = NULL,
  ...
)
}
\arguments{
\item{x}{The \code{durationy} to print or format.}

\item{include_plus}{Whether to include a plus ('+') sign for positive
durations.
Defaults to \code{FALSE}.}

\item{use_true_minus}{Whether to use the
\href{https://www.compart.com/en/unicode/U+2212}{true minus sign ('−', U+2212)}
sign as opposed to the
\href{https://www.compart.com/en/unicode/U+002D}{ASCII hyphen (-, U+2212)}.
Defaults to \code{FALSE}.}

\item{year_unit}{The year unit name to print.
If not blank then the value is followed by a space and the unit.
Cannot be more than 20 characters (UTF-8 bytes) or contain control characters.
Defaults to \code{"yr"}.}

\item{...}{Further arguments to be passed from or to other methods.}

\item{max}{Numeric or \code{NULL}, specifying the maximal number of entries to be
printed. When \code{NULL}, \code{getOption("max.print")} used. Defaults to \code{NULL}.}
}
\description{
A \code{durationy} is printed as a decimal.

This format is readable by \code{\link[=as_durationy.character]{as_durationy.character()}}.
}
\examples{
  pos <- as_durationy(1)
  neg <- as_durationy(-2.3)
  format(pos) # "1 yr"
  format(pos, include_plus = TRUE) # "1 yr"
  format(pos, year_unit = "") # "1"
  format(neg) # "−2.3 yr"
  format(neg, use_true_minus = TRUE) # "−2.3 yr"
  format(neg, year_unit = "a") # "-2.3 a"
}
