% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{integer_constants}
\alias{integer_constants}
\alias{valid_years_start}
\alias{valid_years_end}
\alias{valid_duration_years_max}
\title{Integer constants}
\usage{
valid_years_start

valid_years_end

valid_duration_years_max
}
\description{
The following integer constants may make code clearer.\tabular{lll}{
   Constant \tab Value \tab Meaning \cr
   \code{valid_years_start} \tab \code{1000L} \tab The first calendar year for a \code{datey} \cr
   \code{valid_years_end} \tab \code{3000L} \tab The final valid calendar year for a \code{datey} (noting that only the start of this year is valid) \cr
   \code{valid_duration_years_max} \tab \code{2000L} \tab The maximum valid duration in years for a \code{durationy} \cr
}
}
\examples{
  datey(valid_years_start - 0.001, strict = FALSE)
  datey(valid_years_start)
  datey(valid_years_end)
  datey(valid_years_end + 0.001, strict = FALSE)
  durationy(-(valid_duration_years_max + 0.001), strict = FALSE)
}
\seealso{
\link{is_NA}, \link{NAs}, \link{datey}, \link{all_of_time}
}
