% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datey_interval.R
\name{interval_includes}
\alias{interval_includes}
\alias{\%includes\%}
\title{\code{all_proper(x)} tests whether all the elements of \code{x} are proper.
\code{all_collapsed(x)} tests whether all the elements of \code{x} are collapsed.
\code{any_collapsed(x)} tests whether at least one of the elements of \code{x} is collapsed.
Whether a \code{datey_interval} includes a \code{datey}}
\usage{
interval_includes(interval, value)

interval \%includes\% value
}
\arguments{
\item{interval}{The \code{datey_interval}.}

\item{value}{The \code{datey} to test for inclusion.}
}
\description{
Test whether a \code{datey_interval}, [start, end) includes a \code{datey}, i.e.
start ≤ value and value < end.

The \verb{\%includes\%} operator is syntactic sugar for \code{interval_includes()}.

An NA interval is treated as empty and an NA date is treated as not being in
any interval, so these methods are guaranteed to return \code{TRUE} or \code{FALSE}.
}
