% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datey_interval.R
\name{interval_properties}
\alias{interval_properties}
\alias{$.datey_interval}
\title{Get the start, end or duration of a \code{datey_interval}}
\usage{
\method{$}{datey_interval}(x, name)
}
\arguments{
\item{x}{The \code{datey_interval}.}

\item{name}{Must be \code{start}, \code{end} or \code{duration}.}
}
\value{
\code{start} and \code{end} return a vector of \code{datey};
\code{duration} returns a vector of \code{durationy}.
}
\description{
Get the start, end or duration of a \code{datey_interval} using the syntax
\verb{$start}, \verb{$end} or \verb{$duration} respectively.
}
\examples{
  t_1 <- start_day(2001, 1, 1)
  t_2 <- start_day(2002, 2, 2)
  interval <- datey_interval(t_1, t_2)
  interval
  interval$start
  interval$end
  interval$duration
}
