% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datey_interval.R
\name{interval_start_end}
\alias{interval_start_end}
\alias{$.datey_interval}
\title{Extract start or end from a \code{datey_interval}}
\usage{
\method{$}{datey_interval}(x, name)
}
\arguments{
\item{x}{The \code{datey_interval}.}

\item{name}{Must be \code{start} or \code{end}.}
}
\description{
Extract the start or end of a \code{datey_interval} using the syntax
\verb{$start} or \verb{$end} respectively.
}
\examples{
  t_1 <- start_day(2001, 1, 1)
  t_2 <- start_day(2002, 2, 2)
  interval <- datey_interval(t_1, t_2)
  interval$start
  interval$end
}
