% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/durationy.R
\name{is_durationy}
\alias{is_durationy}
\title{Check if an object is a \code{durationy}}
\usage{
is_durationy(x)
}
\arguments{
\item{x}{The object to test.}
}
\description{
Checks whether an object is a \code{durationy}.
}
