% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datey.R
\name{is_leap_year}
\alias{is_leap_year}
\title{Is year a leap year?}
\usage{
is_leap_year(year)
}
\arguments{
\item{year}{The year to test.}
}
\value{
\code{NA} if \code{year} is non-numeric, non-integral or outside [1000,3000),
\code{TRUE} if \code{year} is a leap year, otherwise
\code{FALSE}.
}
\description{
Tests whether a year is a leap year, \emph{and} that it is
integral and in [1000,3000).
}
\examples{
is_leap_year(2009) # FALSE
is_leap_year(2008) # TRUE
is_leap_year(1900) # FALSE
is_leap_year(2000) # TRUE
}
