% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datey.R, R/durationy.R
\name{is_NA}
\alias{is_NA}
\alias{is.na.datey}
\alias{anyNA.datey}
\alias{is.na.durationy}
\alias{anyNA.durationy}
\title{Whether \code{datey} or \code{durationy} are NA}
\usage{
\method{is.na}{datey}(x)

\method{anyNA}{datey}(x, recursive = FALSE)

\method{is.na}{durationy}(x)

\method{anyNA}{durationy}(x, recursive = FALSE)
}
\arguments{
\item{x}{The \code{datey} or \code{durationy} to test for NA}

\item{recursive}{Currently required to be \code{FALSE} (the default)}
}
\value{
\itemize{
\item \code{is.na()} returns a vector of logical the same length as \code{x}.
\item \code{anyNA()} always returns \code{TRUE} or \code{FALSE}, never \code{NA} and
never anything other than a single value.
}
}
\description{
Valid datey system ranges:
\itemize{
\item Valid dates are from the start of 1000 to the start of 3000.
\item Valid durations are 2000 years or less in magnitude.
}

Values outside the above ranges are treated as NA.

\code{is.na()} tests whether a \code{datey} or \code{durationy} is NA by element.

\code{anyNA()} tests whether any element of a \code{datey} or \code{durationy} is NA.

For convenience,
\itemize{
\item the constants \link{NA_datey_} and \link{NA_datey_} are the \code{datey} and \code{durationy}
versions of NA respectively, and
\item \href{integer_constants}{integer constants} describing the above valid ranges
are also provided.
}

For performance reasons, intermediate calculations may not check for NAs.
}
\examples{
  t <- c(NA_datey_, datey(2000), datey(999.99, strict = FALSE))
  is.na(t) # c(TRUE, FALSE, TRUE)
  anyNA(t) # TRUE

  d <- c(NA_durationy_, durationy(1.5))
  is.na(d) # c(TRUE, FALSE)
  anyNA(d) # TRUE
}
\seealso{
\link{NA_datey_}, \link{NA_datey_}, \link{integer_constants}
}
\keyword{NA}
