% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datey.R
\name{is_start_day}
\alias{is_start_day}
\alias{is_mid_day}
\title{Test whether a \code{datey} is the start or middle of a day}
\usage{
is_start_day(datey)

is_mid_day(datey)
}
\arguments{
\item{datey}{The (vector of ) \code{datey} to test.}
}
\description{
\code{is_start_day()} tests whether the \code{datey} is the start (or end) of
a day, i.e. the boundary between two days.

\code{is_mid_day()} tests whether the \code{datey} is (exactly) the middle of
a day.
}
