% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datey.R, R/datey_interval.R, R/durationy.R
\name{is_type}
\alias{is_type}
\alias{is_datey}
\alias{is_datey_interval}
\alias{is_durationy}
\title{Is \code{x} a \code{datey}, \code{durationy} or \code{datey_interval}?}
\usage{
is_datey(x)

is_datey_interval(x)

is_durationy(x)
}
\arguments{
\item{x}{The object to test.}
}
\value{
A logical scalar indicating whether \code{x} a \code{datey}, \code{durationy} or
\code{datey_interval} as appropriate.
Always \code{FALSE} or \code{TRUE}; never \code{NULL} or \code{NA}.
}
\description{
These methods will always return a scalar logical \code{TRUE} or \code{FALSE}:
\itemize{
\item \code{is_datey()} tests whether an object is a \code{datey}.
\item \code{is_durationy()} tests whether an object is a \code{durationy}.
\item \code{is_datey_interval()} tests whether an object is a \code{datey_interval}.
}
}
\examples{
t <- datey(2000:2001)
t
is_datey(t)
is_datey(NULL)
is_datey(NA)

d <- durationy(0:2)
d
is_durationy(d)
is_durationy(NULL)
is_durationy(NA)

interval <- datey(2000:2001) \%to\% datey(2001:2002)
interval
is_datey_interval(interval)
is_datey_interval(NULL)
is_datey_interval(NA)
}
