% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datey.R
\name{is_xxx_day}
\alias{is_xxx_day}
\alias{is_start_day}
\alias{is_mid_day}
\title{Is a \code{datey} the start (or end) or middle of a day?}
\usage{
is_start_day(x)

is_mid_day(x)
}
\arguments{
\item{x}{The \code{datey} to test.}
}
\description{
\code{is_start_day()} checks whether \code{x} is the start or end of a day.

\code{is_mid_day()} checks whether \code{x} is the middle of a day.

Note that these properties are \emph{not} necessarily preserved when a duration of
\emph{n} years is added or subtracted.
}
