% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{max_min}
\alias{max_min}
\alias{Summary.datey}
\alias{Summary.durationy}
\title{Minimum, maximum or range of \code{datey} or \code{durationy}}
\usage{
\method{Summary}{datey}(..., na.rm = FALSE)

\method{Summary}{durationy}(..., na.rm = FALSE)
}
\arguments{
\item{...}{The \code{datey} or \code{durationy} arguments.}

\item{na.rm}{A logical (\code{TRUE} or \code{FALSE}) indicating whether NA values should be removed before the computation.}
}
\description{
Minimum, maximum or range of \code{datey} or \code{durationy}
}
\examples{
    t <- datey(2000:2003)
    t
    min(t)
    max(t)
    range(t)
}
