% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datey.R
\docType{data}
\name{NA_datey_}
\alias{NA_datey_}
\title{The \code{datey} version of NA}
\format{
An object of class \code{datey_type} (inherits from \code{datey}) of length 1.
}
\usage{
NA_datey_
}
\description{
The \code{datey} version of NA
}
\examples{
  is_datey(NA_datey_) # TRUE
}
\seealso{
\link{is_NA}, \link{NA_durationy_}, \link{integer_constants}
}
\keyword{datasets}
