% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/durationy.R
\name{NA_durationy_}
\alias{NA_durationy_}
\title{The \code{durationy} version of NA}
\usage{
NA_durationy_
}
\description{
The \code{durationy} version of NA
}
\examples{
  is_durationy(NA_durationy_) # TRUE
}
\seealso{
\link{is_NA}, \link{NA_datey_}, \link{integer_constants}
}
