% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datey.R, R/datey_interval.R, R/durationy.R
\name{NAs}
\alias{NAs}
\alias{NA_datey_}
\alias{NA_datey_interval_}
\alias{NA_durationy_}
\title{The \code{datey}, \code{durationy} and \code{datey_interval} versions of NA}
\usage{
NA_datey_

NA_datey_interval_

NA_durationy_
}
\description{
The \code{datey}, \code{durationy} and \code{datey_interval} versions of NA
}
\examples{
  is_datey(NA_datey_)
  is.na(NA_datey_)
  is_durationy(NA_durationy_)
  is.na(NA_durationy_)
  is_datey(NA_datey_interval_)
  is_datey_interval(NA_datey_interval_)
}
\seealso{
\link{is_NA}, \link{integer_constants}
}
