% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{Ops.datey_type}
\alias{Ops.datey_type}
\title{Generic operators for \code{datey}}
\usage{
\method{Ops}{datey_type}(e1, e2)
}
\arguments{
\item{e1}{First (\code{datey}) parameter.}

\item{e2}{Second parameter -- must be \code{datey} or \code{durationy}.}
}
\description{
Generic operators for \code{datey}
}
