% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{ops}
\alias{ops}
\alias{Ops.datey_type}
\title{Operators for \code{datey}, \code{durationy} and \code{datey_interval}}
\usage{
\method{Ops}{datey_type}(e1, e2)
}
\arguments{
\item{e1}{First parameter.}

\item{e2}{Second parameter (missing if a unary operator).}
}
\value{
See above table. In essence
\itemize{
\item subtracting two \code{datey}s results in a \code{durationy},
\item comparing two \code{T}s results in a logical,
\item adding or subtracting a \code{durationy} to or from a \code{T} results in a
\code{T}, and
\item mixing \code{durationy} and \code{datey} with numeric operands first converts
the \code{durationy} and \code{datey} to years and then results in standard numeric
evaluation,
}

where \code{T} is either \code{datey} or \code{durationy} in each of the above.
}
\description{
The unary \code{-} operator can be applied to a \code{durationy} to change its sign.

The following are the available binary operations on \code{datey} and \code{durationy}
only operands, and their meaning:\tabular{ccccl}{
   Left \tab Operators \tab Right \tab Result \tab Notes \cr
   \code{datey} \tab \code{==} \code{!=} \code{<} \code{<=} \code{>} \code{>=} \tab \code{datey} \tab logical \tab Order relation for dates \cr
   \code{datey} \tab \code{-} \tab \code{datey} \tab \code{durationy} \tab Duration between two dates \cr
   \code{datey} \tab \code{+} \code{-} \tab \code{durationy} \tab \code{datey} \tab A date offset by a duration \cr
   \code{durationy} \tab \code{+} \tab \code{datey} \tab \code{datey} \tab A date offset by a duration \cr
   \code{durationy} \tab \code{==} \code{!=} \code{<} \code{<=} \code{>} \code{>=} \tab \code{durationy} \tab logical \tab Order relation for durations \cr
   \code{durationy} \tab \code{+} \code{-} \tab \code{durationy} \tab \code{durationy} \tab Duration addition and substraction \cr
   \code{datey} \tab \code{+} \code{-} \code{==} \code{!=} \code{<} \code{<=} \code{>} \code{>=} \tab numeric \tab numeric \tab The \code{datey} is first converted to years \cr
   numeric \tab \code{+} \code{-} \code{==} \code{!=} \code{<} \code{<=} \code{>} \code{>=} \tab \code{datey} \tab numeric \tab The \code{datey} is first converted to years \cr
   \code{durationy} \tab \code{+} \code{-} \code{*} \code{/} \code{==} \code{!=} \code{<} \code{<=} \code{>} \code{>=} \tab numeric \tab \code{durationy} \tab The \code{durationy} is first converted to years \cr
   numeric \tab \code{+} \code{-} \code{*} \code{/} \code{==} \code{!=} \code{<} \code{<=} \code{>} \code{>=} \tab \code{durationy} \tab \code{durationy} \tab The \code{durationy} is first converted to years \cr
   \code{datey} \tab \verb{\%to\%} \tab \code{datey} \tab \code{datey_interval} \tab Syntactic sugar for \code{datey_interval()} \cr
   \code{datey_interval} \tab \verb{\%includes\%} \tab \code{datey} \tab logical \tab Whether the interval contains the date \cr
}
}
\examples{
t_2000 <- datey(2000)
t_2001 <- datey(2001)
d_0.5 <- durationy(0.5)

t_2000
t_2001
d_0.5

t_2001 - t_2000 # `datey` - `datey` is a `durationy`
t_2000 + d_0.5  # `datey` + `durationy` is a `datey`
t_2001 - d_0.5  # `datey` - `durationy` is a `datey`
t_2000 + 0.5    # Arithmetic with numerics results in a double
d_0.5 + d_0.5   # `durationy` + `durationy` is a `durationy`
d_0.5 + 0.5     # Arithmetic with numerics results in a double
d_0.5 * 2       # Arithmetic with numerics results in a double

interval <- t_2000 \%to\% t_2001
interval
interval \%includes\% t_2000 # TRUE -- start *is* included in an interval
interval \%includes\% (t_2000 + d_0.5) # TRUE
interval \%includes\% t_2001 # FALSE -- end is *not* included in an interval
}
