% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{replicate}
\alias{replicate}
\alias{rep.datey}
\alias{rep.durationy}
\alias{rep.datey_interval}
\title{Replicate \code{datey}, \code{durationy} or \code{datey_interval} vectors}
\usage{
\method{rep}{datey}(x, ...)

\method{rep}{durationy}(x, ...)

\method{rep}{datey_interval}(x, ...)
}
\arguments{
\item{x}{A \code{datey}, \code{durationy} or \code{datey_interval}.}

\item{...}{Other arguments.}
}
\value{
The replicated vector.
}
\description{
Replicates \code{datey}, \code{durationy} or \code{datey_interval} vectors.
}
\examples{
  x <- datey(2001:2004)
  rep(x, 2)
  rep(x, each = 2)
  pmax(x, datey(2002))
}
\seealso{
\link{datey}, \link{durationy}, \link{datey_interval}
}
\keyword{replicate}
