% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{seq}
\alias{seq}
\alias{seq.datey}
\alias{seq.durationy}
\title{Create \code{datey} or \code{durationy} sequence vector}
\usage{
\method{seq}{datey}(from, to, by)

\method{seq}{durationy}(from, to, by)
}
\arguments{
\item{from, to}{The starting and (maximal) end values of the sequence.
Scalar \code{datey}s or \code{durationy}s.}

\item{by}{A scalar\code{durationy} representing the increment of the sequence.}
}
\value{
The sequence
}
\description{
Creates a \code{datey} or \code{durationy} vector by defining a sequence.
}
\examples{
  seq(from = datey(2000), to = datey(2005), by = durationy(2))
  # 2000-01-01 2002-01-01 2004-01-01
}
\keyword{sequence}
