% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{subset}
\alias{subset}
\alias{[.datey}
\alias{[.durationy}
\alias{[<-.datey}
\alias{[<-.durationy}
\title{Subset \code{datey} or \code{durationy} vectors}
\usage{
\method{[}{datey}(x, i, ...)

\method{[}{durationy}(x, i, ...)

\method{[}{datey}(x, i) <- value

\method{[}{durationy}(x, i) <- value
}
\arguments{
\item{x}{A \code{datey} or \code{durationy}.}

\item{i}{Indices to extract.}

\item{...}{Passed through.}

\item{value}{Value to assign.}
}
\value{
The subset.
}
\description{
Subsets \code{datey} or \code{durationy} vectors.
}
\examples{
  x <- datey(2001:2004) # 2001-01-01 2002-01-01 2003-01-01 2004-01-01
  x[2:3] # 2002-01-01 2003-01-01
  x[2:3] <- datey(1999) # 2001-01-01 1999-01-01 1999-01-01 2004-01-01

}
\keyword{classes}
\keyword{subset}
