% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datey.R
\name{text_from_datey}
\alias{text_from_datey}
\alias{as.character.datey}
\alias{format.datey}
\alias{print.datey}
\title{Format or print a \code{datey}}
\usage{
\method{as.character}{datey}(x, ...)

\method{format}{datey}(x, include_day_fraction = TRUE, ...)

\method{print}{datey}(x, include_day_fraction = TRUE, max = NULL, ...)
}
\arguments{
\item{x}{The \code{datey} to print or format.}

\item{...}{Other arguments.}

\item{include_day_fraction}{Whether to include the fractional day part.
Defaults to \code{TRUE}.}

\item{max}{Numeric or \code{NULL}, specifying the maximal number of entries to be
printed. When \code{NULL}, \code{getOption("max.print")} used. Defaults to \code{NULL}.}
}
\value{
\code{as.character} and \code{format} return a vector of \code{character}.
\code{print} invisibly returns \code{x}.
}
\description{
A \code{datey} is printed as either
\itemize{
\item \code{YYYY-MM-DD}, i.e. ISO 8601 extended date format, or
\item \code{YYYY-MM-DD.F...} where \code{.F...} is the day fraction part.
}

If \code{include_day_fraction} is \code{TRUE} then \verb{[.F...]} is included even if it
is 0 (i.e. \code{.0}).

Note that a \code{datey} created as the end of a day (or with day fraction 1) will
print as the start of the following day.
}
\examples{
start      <- start_day(2001, 2, 3)
fractional <- datey(    2001, 2, 3, day_fraction = 0.4444)
mid        <- mid_day(  2001, 2, 3)
end        <- end_day(  2001, 2, 3)

format(start)                               # "2001-02-03.0"
format(start, include_day_fraction = FALSE) # "2001-02-03"
format(fractional)                          # "2001-02-03.4447"
format(mid)                                 # "2001-02-03.5"
format(end)                                 # "2001-02-04.0"
}
\seealso{
\link{datey}
}
