% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datey_interval.R
\name{text_from_datey_interval}
\alias{text_from_datey_interval}
\alias{as.character.datey_interval}
\alias{format.datey_interval}
\alias{print.datey_interval}
\title{Format or print a \code{datey_interval}}
\usage{
\method{as.character}{datey_interval}(x, ...)

\method{format}{datey_interval}(x, include_day_fraction = TRUE, ...)

\method{print}{datey_interval}(x, include_day_fraction = TRUE, max = NULL, ...)
}
\arguments{
\item{x}{The \code{datey_interval} to print or format.}

\item{...}{Further arguments to be passed from or to other methods.}

\item{include_day_fraction}{Whether to include the fractional day part.
Defaults to \code{TRUE}.}

\item{max}{Numeric or \code{NULL}, specifying the maximal number of entries to be
printed. When \code{NULL}, \code{getOption("max.print")} used. Defaults to \code{NULL}.}
}
\value{
\code{as.character} and \code{format} return a vector of \code{character}.
\code{print} invisibly returns \code{x}.
}
\description{
A \code{datey_interval} is printed as "[start, end)", where start and end are
printed either as
\itemize{
\item \code{YYYY-MM-DD}, i.e. ISO 8601 extended date format, or
\item \code{YYYY-MM-DD.F...} where \code{.F...} is the day fraction part.
}

If \code{include_day_fraction} is \code{TRUE} then \verb{[.F...]} is included even if it
is 0 (i.e. \code{.0}).
}
\seealso{
\link{datey_interval}
}
