% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/durationy.R
\name{text_from_durationy}
\alias{text_from_durationy}
\alias{as.character.durationy}
\alias{format.durationy}
\alias{print.durationy}
\title{Format or print a \code{durationy}}
\usage{
\method{as.character}{durationy}(x, ...)

\method{format}{durationy}(x, include_plus = FALSE, use_true_minus = TRUE, year_unit = "yr", ...)

\method{print}{durationy}(
  x,
  include_plus = FALSE,
  use_true_minus = TRUE,
  year_unit = "yr",
  max = NULL,
  ...
)
}
\arguments{
\item{x}{The \code{durationy} to print or format.}

\item{...}{Other arguments.}

\item{include_plus}{Whether to include a plus ('+') sign for positive
durations.
Defaults to \code{FALSE}.}

\item{use_true_minus}{Whether to use the
\href{https://www.compart.com/en/unicode/U+2212}{true minus sign ('-', U+2212)}
sign as opposed to the
\href{https://www.compart.com/en/unicode/U+002D}{ASCII hyphen (-, U+002D)}.
Defaults to \code{TRUE}.}

\item{year_unit}{The year unit name to print.
If not blank then the value is followed by a space and the unit.
Cannot be more than 20 characters (UTF-8 bytes) or contain control characters.
Defaults to \code{"yr"}.}

\item{max}{Numeric or \code{NULL}, specifying the maximal number of entries to be
printed. When \code{NULL}, \code{getOption("max.print")} used. Defaults to \code{NULL}.}
}
\value{
\code{as.character} and \code{format} return a vector of \code{character}.
\code{print} invisibly returns \code{x}.
}
\description{
A \code{durationy} is printed as a decimal.

This format is readable by \code{\link[=durationy.character]{durationy.character()}}.
}
\examples{
pos <- durationy(1)
neg <- durationy(-2.3)
format(pos) # "1 yr"
format(pos, include_plus = TRUE) # "+1 yr"
format(pos, year_unit = "") # "1"
format(neg) # U+2212 (true minus) followed by "2.3" (CRAN-compliance)
format(neg, use_true_minus = FALSE) # "-2.3 yr"
format(neg, use_true_minus = FALSE, year_unit = "a") # "-2.3 a"
}
\seealso{
\link{durationy}
}
