% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/durationy.R
\name{text_to_durationy}
\alias{text_to_durationy}
\alias{durationy.character}
\title{Parse text as a \code{durationy}}
\usage{
\method{durationy}{character}(x, strict = TRUE, blank_is_NA = FALSE, year_unit = "yr", ...)
}
\arguments{
\item{x}{Vector of text items to be parsed.}

\item{strict}{How non-compliant text (including values greater than
2000 in magnitude) should be handled.
If \code{strict} is \code{TRUE} then execution is stopped.
If \code{strict} is \code{FALSE} then \code{NA} is returned.
Defaults to \code{TRUE}.}

\item{blank_is_NA}{Whether blanks should be treated as \code{NA}.
Defaults to \code{FALSE}.}

\item{year_unit}{The year unit name to expect.
If not blank then the value is expected to be followed by a space and this unit text.
Cannot be more than 20 characters (UTF-8 bytes) or contain control
characters.
Defaults to \code{"yr"}.}

\item{...}{Other arguments (not used in this package).}
}
\value{
A vector of \code{durationy}.
}
\description{
This function parses text  a \code{durationy}.

If the text is NA then NA is returned.

Valid text is of the form "DDD.DDD" or "DDD.DDD UUU"
where
\itemize{
\item "DDD.DDD" is the duration in years with "D" being decimal digits
(and the fractional part is not required for whole years), and
\item "UUU" is the unit text (which defaults to "yrs"). If blank then there is
no space after the duration in years.
}

If \code{blank_is_NA} is \code{TRUE} then blanks are treated as \code{NA}.

If \code{strict} is \code{TRUE} (which is the default) then non-compliant text will
stop execution.
}
\examples{
durationy("10 yr")
durationy("+10 yr")
durationy("-10 yr")
durationy("10", year_unit = "")
durationy("10 a", year_unit = "a")

# Handling blanks:
try(durationy(""))
durationy("", blank_is_NA = TRUE)

# Invalids:
try(durationy("abc"))
try(durationy("2000.000001 yr"))
durationy("abc", strict = FALSE) # NA
durationy("2000.000001 yr", strict = FALSE) # NA
durationy("2000.000000 yr") # This is valid
}
