% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datey.R
\docType{data}
\name{valid_years_end}
\alias{valid_years_end}
\title{The first \emph{invalid} calendar year after \link{valid_years_start}.}
\format{
An object of class \code{integer} of length 1.
}
\usage{
valid_years_end
}
\description{
The first \emph{invalid} calendar year after \link{valid_years_start}.
}
\keyword{datasets}
