% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datey.R
\docType{data}
\name{valid_years_start}
\alias{valid_years_start}
\title{The first \emph{valid} calendar year.}
\format{
An object of class \code{integer} of length 1.
}
\usage{
valid_years_start
}
\description{
The first \emph{valid} calendar year.
}
\keyword{datasets}
