% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datey.R
\name{ymdf}
\alias{ymdf}
\alias{to_ymdf}
\alias{point_in_day}
\title{\code{datey} in terms of calendar year, month, day and day fraction components}
\usage{
to_ymdf(datey)

point_in_day(year, month, day, day_fraction, strict = TRUE)
}
\arguments{
\item{datey}{The \code{datey} to be deconstructed by \code{from_ymdf()}.}

\item{year}{Calendar year.
Valid years are from 1000 (inclusive) to 3000 (exclusive).
If provided as \code{double} then these \emph{must be integers}.}

\item{month}{Month number in calendar year, with 1 representing January.
If provided as \code{double} then these \emph{must be integers}.}

\item{day}{Day number in month, with 1 representing the first day of the month.
If provided as \code{double} then these \emph{must be integers}.}

\item{day_fraction}{The fraction of the day, in [0,1].
0 means the start of the day,
0.5 means the middle of the day, and
1 means the end of the day
(which is identical to the start of the next day).}

\item{strict}{How to handle calendar years less than 1000 or greater than or equal to 3000 and day
fractions not in the interval [0,1].
\itemize{
\item If \code{strict} is \code{TRUE} -- the default -- then execution is stopped.
\item If \code{strict} is \code{FALSE} then \code{NA} is returned.
}

(NA arguments result in NA regardless of \code{strict}.)}
}
\description{
\code{to_ymdf()} returns a list of the \code{year}, \code{month}, \code{day} and \code{day_fraction}
breakdown of a \code{datey}, where
\itemize{
\item \code{year} is an \code{integer} in [1000,3000),
\item \code{month} is an \code{integer} in [1,12],
\item \code{day} is an \code{integer} in [1,N], where N is the number of days
in the month specified by \code{year} and \code{month}, and
\item \code{day_fraction} is a \code{double} in [0,1) representing the fraction of the day,
where e.g. 0 means the start and 0.5 means the middle of the day.
}

If the \code{datey} was constructed using \code{end_day} or \code{day_fraction = 1} then
\code{to_ymdf()} will return the \emph{start} of the \emph{next} day with \code{day_fraction = 0}.

\code{from_ymdf()} creates a \code{datey} from
a calendar year, month, and day fraction.

The lengths of vector arguments must be multiples of each other.
}
\section{Edge cases}{
The following are special cases:
\enumerate{
\item \code{from_ymdf(999, 12, 31, 1)} \emph{will} cause an error,
even though in theory it represents the legal \code{datey} 1000-01-01.0.
\item \code{from_ymdf(2999, 12, 31, 1)} will \emph{not} cause an error,
even though in theory it represents the illegal \code{datey}
3000-01-01.0. The resulting \code{datey} will be invalid though.
}
}

