% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datey_interval.R
\name{all_of_time}
\alias{all_of_time}
\title{Maximum valid \code{datey_interval}}
\usage{
all_of_time
}
\description{
The \code{datey_interval} [1000-01-01.0, 3000-01-01.0).

This is the value used when a \code{datey_interval} is combined with a logical
using the \code{&} operator.
}
\examples{
  all_of_time
}
