% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datey.R
\name{as_years_datey}
\alias{as_years_datey}
\alias{as.double.datey}
\alias{as.integer.datey}
\title{Convert a \code{datey} to calendar years (including fractional part)}
\usage{
\method{as.double}{datey}(x, ...)

\method{as.integer}{datey}(x, ...)
}
\arguments{
\item{x}{The \code{datey} to convert to years.}

\item{...}{Other arguments (not used in this package).}
}
\description{
Converts a \code{datey} to calendar years. The fractional part represents the
proportion of the calendar year that has elapsed.

For example,
\itemize{
\item the \emph{start} of 2000-01-01 (or, equivalently, the \emph{end} of 1999-12-31),
results in \code{2000}, and
\item the \emph{middle} of the calendar year 2000 results in \code{2000.5}.
}

Note the following:
\itemize{
\item \code{as.numeric()} is the same as \code{as.double()}.
\item \code{as.integer()} gives the calendar year, e.g.
\code{as.integer(datey(2000.9))} is \code{2000}.
\code{as.integer(x)} is the same as \code{as.integer(as.double(x))}.
}
}
