% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/durationy.R
\name{as_years_durationy}
\alias{as_years_durationy}
\alias{as.double.durationy}
\alias{as.integer.durationy}
\title{Convert a \code{durationy} to duration in years}
\usage{
\method{as.double}{durationy}(x, ...)

\method{as.integer}{durationy}(x, ...)
}
\arguments{
\item{x}{The \code{durationy} to convert to years.}

\item{...}{Other arguments (not used in this package).}
}
\description{
Converts a \code{durationy} to duration in years.

Note the following:
\itemize{
\item \code{as.numeric()} is the same as \code{as.double()}.
\item \code{as.integer()} obtains the integer part,
e.g. \code{as.integer(durationy(1.9))} is \code{1}
and \code{as.integer(durationy(-1.9))} is \code{-1}.
\code{as.integer(x)} is the same as \code{as.integer(as.double(x))}.
}
}
