% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datey-package.R
\docType{package}
\name{datey-package}
\alias{datey-package}
\title{datey: Exact Date and Duration Arithmetic on an Annual Grid}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Standardised mapping of dates onto a discrete annual grid, together with exact date and duration arithmetic. This matters when the primary unit is years but the input data uses dates. Examples are actuarial mortality experience analysis and valuation of life assurance and annuities, for which mortality rates are defined per year but experience and valuation data are typically defined using dates.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/logmu-org/r-datey}
  \item \url{https://r-datey.logmu.org/}
  \item Report bugs at \url{https://github.com/logmu-org/r-datey/issues}
}

}
\author{
\strong{Maintainer}: Tim Gordon \email{tim.gordon@btinternet.com} (\href{https://orcid.org/0009-0006-3087-4384}{ORCID})

Authors:
\itemize{
  \item Tim Gordon \email{tim.gordon@btinternet.com} (\href{https://orcid.org/0009-0006-3087-4384}{ORCID})
}

}
\keyword{internal}
