% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datey-package.R
\docType{package}
\name{datey-package}
\alias{datey-package}
\title{datey: Annualised Dates and Durations for Actuarial Analytics}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Tim Gordon \email{tim.gordon@btinternet.com} (\href{https://orcid.org/0009-0006-3087-4384}{ORCID})

}
\keyword{internal}
