% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datey-package.R
\docType{package}
\name{datey-package}
\alias{datey-package}
\title{datey: Annualised Fixed-Precision Dates and Durations for Actuarial Experience Analysis}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

TODO.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/logmu-org/r-datey}
  \item \url{http://logmu.org/r-datey/}
  \item \url{http://logmu-org.github.io/r-datey/}
  \item \url{https://logmu-org.github.io/r-datey/}
}

}
\author{
\strong{Maintainer}: Tim Gordon \email{tim.gordon@btinternet.com} (\href{https://orcid.org/0009-0006-3087-4384}{ORCID})

}
\keyword{internal}
