% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datey-package.R
\docType{package}
\name{datey-package}
\alias{datey-package}
\title{datey: Date and Duration Arithmetic on an Annual Grid}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A standardised mapping of dates onto an annual grid with fixed precision and associative date and duration-related arithmetic. This matters when the primary unit is years but the input data uses dates. Examples are actuarial mortality experience analysis or valuation of life assurance and annuities, for which mortality rates are defined per year but experience and valuation data use dates.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/logmu-org/r-datey}
  \item \url{https://logmu.org/r-datey/}
  \item \url{https://logmu-org.github.io/r-datey/}
  \item Report bugs at \url{https://github.com/logmu-org/r-datey/issues}
}

}
\author{
\strong{Maintainer}: Tim Gordon \email{tim.gordon@btinternet.com} (\href{https://orcid.org/0009-0006-3087-4384}{ORCID})

Authors:
\itemize{
  \item Tim Gordon \email{tim.gordon@btinternet.com} (\href{https://orcid.org/0009-0006-3087-4384}{ORCID})
}

}
\keyword{internal}
