% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datey.R
\name{datey}
\alias{datey}
\alias{as_start_day}
\alias{as_mid_day}
\alias{as_end_day}
\title{Create a \code{datey}}
\usage{
datey(year, month, day, day_fraction)

as_start_day(x, ...)

as_mid_day(x, ...)

as_end_day(x, ...)
}
\arguments{
\item{year}{Calendar year.
Valid years are from 1000 (inclusive) to 3000 (exclusive).}

\item{month}{Month number in calendar year, with 1 representing January.}

\item{day}{Day number in month, with 1 representing the first day of the month.}

\item{day_fraction}{The fraction of the day, in [0,1].
0 means the start of the day,
0.5 means the middle of the day, and
1 means the end of the day
(which is identical to the start of the next day).}
}
\description{
Creates a \code{datey} from
a year, month, day and day-fraction.
}
\details{
\code{year}, \code{month} and \code{day} must be integral and of the
same length.

\code{day_fraction} must be numeric, lie in [0,1]
and be either the same length as the date arguments or a scalar.

Calendar years outside the interval [1000,3000)
are treated as NA.

To deconstruct a \code{datey}, use \link{as_ymdf}.
}
