% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datey.R
\name{datey}
\alias{datey}
\alias{datey.default}
\alias{datey.datey}
\alias{datey.integer}
\alias{datey.double}
\alias{datey.Date}
\alias{datey.POSIXct}
\alias{datey.POSIXlt}
\title{Create a \code{datey} from a calendar year (including its fractional part) or
another date type}
\usage{
datey(x, day_fraction = NULL, strict = TRUE, ...)

\method{datey}{default}(x, day_fraction = NULL, strict = TRUE, ...)

\method{datey}{datey}(x, day_fraction = NULL, strict = TRUE, ...)

\method{datey}{integer}(x, day_fraction = NULL, strict = TRUE, ...)

\method{datey}{double}(x, day_fraction = NULL, strict = TRUE, ...)

\method{datey}{Date}(x, day_fraction = NULL, strict = TRUE, ...)

\method{datey}{POSIXct}(x, day_fraction = NULL, strict = TRUE, ...)

\method{datey}{POSIXlt}(x, day_fraction = NULL, strict = TRUE, ...)
}
\arguments{
\item{x}{The argument to convert to a \code{datey}.}

\item{day_fraction}{The \code{day_fraction} override. Defaults to \code{NULL}.

If \code{day_fraction} is \emph{not} provided then \code{x} is used to derive both
the calendar year, month, day \emph{and} the day fraction.

If \code{day_fraction} \emph{is} provided then \code{x} is used solely to derive the
calendar year, month and day, while \code{day_fraction} provides the position in
the day. \code{day_fraction} must lie in the inclusive interval [0,1], with
0 meaning the start of the day,
0.5 meaning the middle of the day, and
1 meaning the end of the day (which is identical to the start of the next
day).}

\item{strict}{How calendar years less than 1000 or greater than 3000 and day
fractions not in the interval [0,1] should be
handled.
If \code{strict} is \code{TRUE} -- the default -- then execution is stopped.
If \code{strict} is \code{FALSE} then \code{NA} is returned.

NA arguments result in NA (and do not stop execution) regardless of \code{strict}.}

\item{...}{Other arguments (not used in this package).}
}
\value{
A vector of \code{datey}.
}
\description{
This package provides methods to create a \code{datey} from the following:
\itemize{
\item \code{double} and \code{integer} are interpreted as the specified
calendar year, with the fractional part representing the fraction of the
year. For instance, \code{datey(2000.5)} means halfway through the year 2000.
(This means that an \code{integer} argument always indicates the \emph{start} of the
calendar year, e.g. \code{datey(2000L)} is the start of the year 2000.)
\item \code{Date} and \code{POSIXct} and \code{POSIXlt} are interpreted as fractional
years. If no \code{day_fraction} argument is provided then the day fraction is
determined by the hours, minutes, and seconds. For instance,
\code{datey(as.POSIXct("2000-03-21 12:00"))} means the \emph{middle} of 2000-03-21.
Note that in standard use, a \code{Date} has no fractional part and therefore
means the \emph{start} of the day. For instance,
\code{datey(as.Date("2000-03-21 12:00"))} means the \emph{start} of 2000-03-21.
\item \code{character} -- If \code{day_fraction} \emph{is} provided then the text format must
be in ISO 8601 extended format, i.e. YYYY-MM-DD. If \code{day_fraction} is \emph{not}
provided then the text format must be YYYY-MM-DD.FFF, where .FFF is the day
fraction and must be present even if the fraction is 0, e.g. "2000-01-01.0"
for the start of 1 January 2000.
\item \code{datey} is interpreted as is but with the optional \code{day_fraction} override.
Note that a \code{day_fraction} of 1 will add a day to a day boundary,
\emph{even if it was originally defined as an end day}.
}

The lengths of vector arguments must be multiples of each other.

This is an S3 generic.
}
\examples{
datey(2000)
datey(2000.5) # Middle of a leap year
datey(2001.5) # Middle of a non-leap year
datey(as.Date("2020-01-02"))
datey(as.POSIXct("2020-01-02 12:00:00"))
datey(as.POSIXlt("2020-01-02 12:00:00"))

# Use `strict` to control error behaviour for invalid `datey`s:
try(datey(999.9))
try(datey(3000.1))
datey(999.9, strict = FALSE)
datey(3000.1, strict = FALSE)
}
\seealso{
Use \code{\link[=start_day]{start_day()}}, \code{\link[=mid_day]{mid_day()}} and \code{\link[=end_day]{end_day()}} to create a \code{datey}
direct from year, month and day. Use \code{\link[=as_start_day]{as_start_day()}}, \code{\link[=as_mid_day]{as_mid_day()}} and
\code{\link[=as_end_day]{as_end_day()}} to create a \code{datey} from a numeric or base R date type but
specifying whether it should be the start, middle or end of the day.

See \link{text_to_datey} for parsing and creating \code{datey} text.
}
