% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datey_interval.R
\name{datey_interval}
\alias{datey_interval}
\alias{\%to\%}
\title{Create a \code{datey_interval}}
\usage{
datey_interval(start, end, strict = TRUE)

start \%to\% end
}
\arguments{
\item{start, end}{The start (inclusive) and end of the interval (exclusive).
These can be any type that is convertible to a \code{datey}. Must have the same
numbers of elements or be multiples of each other.}

\item{strict}{How NAs should be handled.
\itemize{
\item If \code{strict} is \code{TRUE} -- the default -- then execution is stopped.
\item If \code{strict} is \code{FALSE} then \code{NA} is returned if \code{start} and/or \code{end} is NA.
}}
}
\description{
Creates the \code{datey_interval} [\code{start}, \code{end}).

The operator syntax may be clearer, i.e. \code{start \%to\% end}.
}
\examples{
  t_1 <- start_day(2001, 1, 1)
  t_2 <- start_day(2002, 2, 2)
  datey_interval(t_1, t_2)
  t_1 \%to\% t_2
}
