% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datey_interval.R
\name{datey_interval}
\alias{datey_interval}
\alias{\%to\%}
\title{Create a \code{datey_interval}}
\usage{
datey_interval(start, end, strict = TRUE)

start \%to\% end
}
\arguments{
\item{start, end}{The start (inclusive) and end of the interval (exclusive).
These can be any type that is convertible to a \code{datey}. These have the same
numbers of elements or their lengths must be multiples of each other.}

\item{strict}{How NAs should be handled.
If \code{strict} is \code{TRUE} -- the default -- then execution is stopped.
If \code{strict} is \code{FALSE} then \code{NA} is returned if \code{start} and/or \code{end} is NA.}
}
\value{
A vector of \code{datey_interval}.
}
\description{
Create a \code{datey_interval} representing [\code{start}, \code{end}).

These are closed-open ('clopen') intervals \code{start} <= t < \code{end}, i.e. the
interval includes \code{start} but excludes \code{end}.

There are two syntaxes:
\itemize{
\item operator: \code{start \%to\% end}
\item function: \code{datey_interval(start, end)}
}

These are equivalent other than \code{strict} is always on for the operator
version.
}
\examples{
  datey(1999) \%to\% mid_day(2025, 7, 15)
  datey(1999) \%to\% datey(2000:2002)
}
