% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/durationy.R
\name{durationy}
\alias{durationy}
\alias{durationy.default}
\alias{durationy.durationy}
\alias{durationy.integer}
\alias{durationy.double}
\title{Create a \code{durationy} from an annual duration}
\usage{
durationy(x, strict = TRUE, ...)

\method{durationy}{default}(x, strict = TRUE, ...)

\method{durationy}{durationy}(x, strict = TRUE, ...)

\method{durationy}{integer}(x, strict = TRUE, ...)

\method{durationy}{double}(x, strict = TRUE, ...)
}
\arguments{
\item{x}{The argument to convert to a \code{durationy}.}

\item{strict}{How years greater than 2000 in magnitude should be
handled.
\itemize{
\item If \code{strict} is \code{TRUE} -- the default -- then execution is stopped.
\item If \code{strict} is \code{FALSE} then \code{NA} is returned.
}

(NAs will result in NA regardless of this switch.)}

\item{...}{Other arguments (not used in this package).}
}
\description{
This package provides methods to create a \code{durationy} from the following:
\itemize{
\item \code{integer} -- the value is interpreted as the specified
number of years.
\item \code{double} -- the value is interpreted as the specified
number of years, rounded to fixed precision of a \code{durationy}. This means that
\code{durationy(0.5)} is precise but \code{durationy(0.01)} is not.
\item \code{durationy} -- value is unchanged.
}

This is an S3 generic.
}
