% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{integer_constants}
\alias{integer_constants}
\alias{valid_years_start}
\alias{valid_years_end}
\alias{valid_duration_years_max}
\title{Integer constants}
\usage{
valid_years_start

valid_years_end

valid_duration_years_max
}
\description{
The following integer constants may make code clearer.\tabular{lll}{
   Constant \tab Value \tab Meaning \cr
   \code{valid_years_start} \tab \code{1000L} \tab The first \emph{valid} calendar year for a \code{datey} \cr
   \code{valid_years_end} \tab \code{3000L} \tab The first \emph{invalid} calendar year for a \code{datey} after after \cr
   \code{valid_years_start} \tab  \tab  \cr
   \code{valid_max_duration} \tab \code{2000L} \tab The maximum valid \cr
   duration in years for a \code{durationy} \tab  \tab  \cr
}
}
\examples{
  x <- c(NA_durationy_, durationy(1.5))
  is.na(x) # c(TRUE, FALSE)
}
\seealso{
\link{is_NA}, \link{NA_datey_}, \link{NA_durationy_}
}
\keyword{NA}
