% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datey_interval.R
\name{interval_includes}
\alias{interval_includes}
\alias{\%includes\%}
\title{Whether a \code{datey_interval} includes a \code{datey}}
\usage{
interval_includes(interval, value)

interval \%includes\% value
}
\arguments{
\item{interval}{The \code{datey_interval}.}

\item{value}{The \code{datey} to test for inclusion. A numeric value is also
accepted and is coerced to \code{datey}.}
}
\value{
A vector of \code{logical} corresponding to whether the interval
includes the value. Always TRUE or FALSE -- NAs result in FALSE.
}
\description{
Test whether a \code{datey_interval}, \eqn{[a,b)}, includes a \code{datey} \eqn{t}, i.e.
\eqn{a \le t}{a <= t} and \eqn{t < b}.

The \verb{\%includes\%} operator is syntactic sugar for \code{interval_includes()}.

An NA interval is treated as empty and an NA date is treated as not being in
any interval, so these methods are guaranteed to return \code{TRUE} or \code{FALSE}.
}
\examples{
  t_2000 <- datey(2000)
  t_2001 <- datey(2001)
  t_2002 <- datey(2002)
  t_2003 <- datey(2003)
  t_2004 <- datey(2004)

  interval <- t_2000 \%to\% t_2003
  interval \%includes\% t_2000
  interval \%includes\% t_2001 # Start of interval *is* included
  interval \%includes\% t_2002
  interval \%includes\% t_2003 # End of interval *not* included
  interval \%includes\% t_2004
  interval \%includes\% NA_datey_        # NAs are FALSE
  NA_datey_interval_ \%includes\% t_2004 # NAs are FALSE
  interval_includes(NA_datey_interval_, t_2002) # NAs are FALSE

  # Function syntax:
  interval_includes(interval, t_2002)
}
\seealso{
\link{datey_interval}, \link{interval_properties}, \link{interval_nature}, \link{ops}
}
