% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datey_interval.R
\name{interval_nature}
\alias{interval_nature}
\alias{is_proper}
\alias{is_proper.default}
\alias{is_proper.datey_interval}
\alias{all_proper}
\alias{all_proper.default}
\alias{all_proper.datey_interval}
\alias{is_collapsed}
\alias{is_collapsed.default}
\alias{is_collapsed.datey_interval}
\alias{all_collapsed}
\alias{all_collapsed.default}
\alias{all_collapsed.datey_interval}
\alias{any_collapsed}
\alias{any_collapsed.default}
\alias{any_collapsed.datey_interval}
\title{Properties of a \code{datey_interval}}
\usage{
is_proper(x)

\method{is_proper}{default}(x)

\method{is_proper}{datey_interval}(x)

all_proper(x)

\method{all_proper}{default}(x)

\method{all_proper}{datey_interval}(x)

is_collapsed(x)

\method{is_collapsed}{default}(x)

\method{is_collapsed}{datey_interval}(x)

all_collapsed(x)

\method{all_collapsed}{default}(x)

\method{all_collapsed}{datey_interval}(x)

any_collapsed(x)

\method{any_collapsed}{default}(x)

\method{any_collapsed}{datey_interval}(x)
}
\arguments{
\item{x}{The interval to test.}
}
\value{
\itemize{
\item \code{is_XXX} functions return a logical vector corresponding the property.
\item \code{all_XXX} and \code{any_XXX} functions return a logical scalar.
}
}
\description{
Test whether intervals, \eqn{[a,b)}, are 'proper' or 'collapsed':
\itemize{
\item A \emph{proper} interval does not end before its start,
i.e. \eqn{a \le b}{a <= b}.
\item A \emph{collapsed} interval does not start before its end,
i.e. \eqn{a \ge b}{a >= b}.
}

An NA interval is treated as collapsed and improper.

These definitions imply the following:
\itemize{
\item A collapsed interval could be empty or improper.
\item To test for an empty interval, i.e. \eqn{[a,a)}, test that it is both
proper and collapsed.
}

These methods are guaranteed to return \code{TRUE} or \code{FALSE}, i.e. not \code{NA}
(provided the argument is an interval).

Vector versions mapping each element of \code{x} to \code{TRUE} or \code{FALSE}:

\code{is_proper(x)} tests whether the elements of \code{x} are proper.
\code{is_collapsed(x)} tests whether the elements of \code{x} are collapsed.

Scalar versions mapping \code{x} to a scalar \code{TRUE} or \code{FALSE}:

\code{all_proper(x)} tests whether all the elements of \code{x} are proper.
\code{all_collapsed(x)} tests whether all the elements of \code{x} are collapsed.
\code{any_collapsed(x)} tests whether at least one of the elements of \code{x} is
collapsed.

(\code{any_proper()} is not implemented because there is no obvious use case.)

These are S3 generic functions.
}
\examples{
  non_empty   <- 1900 \%to\% 2000
  empty       <- 2000 \%to\% 2000
  improper    <- 2000 \%to\% 1900
  na          <- NA_datey_interval_

  is_collapsed(non_empty) # FALSE
  is_collapsed(empty)     # TRUE
  is_collapsed(improper)  # TRUE
  is_collapsed(na)        # TRUE

  is_proper(non_empty)    # TRUE
  is_proper(empty)        # TRUE
  is_proper(improper)     # FALSE
  is_proper(na)           # FALSE
}
\seealso{
\link{datey_interval}, \link{interval_properties}, \link{interval_includes}
}
