% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datey.R
\name{is_leap_year}
\alias{is_leap_year}
\alias{is_leap_year.default}
\alias{is_leap_year.integer}
\alias{is_leap_year.double}
\alias{is_leap_year.datey}
\alias{is_leap_year.Date}
\alias{is_leap_year.POSIXct}
\title{Is \code{x} a leap year?}
\usage{
is_leap_year(x, ...)

\method{is_leap_year}{default}(x, ...)

\method{is_leap_year}{integer}(x, ...)

\method{is_leap_year}{double}(x, ...)

\method{is_leap_year}{datey}(x, ...)

\method{is_leap_year}{Date}(x, ...)

\method{is_leap_year}{POSIXct}(x, ...)

\method{is_leap_year}{POSIXct}(x, ...)
}
\arguments{
\item{x}{A vector date type or numeric year.}

\item{...}{Additional arguments (unused by this package).}
}
\value{
\code{NA} if \code{x} is not interpretable as a year or date, or outside [1000,3000),
\code{TRUE} if \code{x} is a leap year, otherwise
\code{FALSE}.
}
\description{
Tests whether a date or year is a leap year.

For years outside [1000,3000), this returns \code{NA}.

This is an S3 generic. This package provides methods for:
\itemize{
\item date types \code{datey}, \code{Date}, \code{POSIXct} and \code{POSIXlt}, and
\item numeric types \code{double} and \code{integer}.
}
}
\examples{
is_leap_year(2009) # FALSE
is_leap_year(2008.5) # TRUE
is_leap_year(datey(1900, 1, 1, 0)) # FALSE
is_leap_year(2000) # TRUE
}
