% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datey.R, R/datey_interval.R, R/durationy.R
\name{is_NA}
\alias{is_NA}
\alias{is.na.datey}
\alias{anyNA.datey}
\alias{is.na.datey_interval}
\alias{anyNA.datey_interval}
\alias{is.na.durationy}
\alias{anyNA.durationy}
\title{Whether \code{datey}, \code{durationy} or
\code{datey_interval} are NA}
\usage{
\method{is.na}{datey}(x)

\method{anyNA}{datey}(x, recursive = FALSE)

\method{is.na}{datey_interval}(x)

\method{anyNA}{datey_interval}(x, recursive = FALSE)

\method{is.na}{durationy}(x)

\method{anyNA}{durationy}(x, recursive = FALSE)
}
\arguments{
\item{x}{The \code{datey}, \code{durationy} or \code{datey_interval} to test for NA.}

\item{recursive}{Currently required to be \code{FALSE} (the default).}
}
\value{
\code{is.na()} returns a vector of logical the same length as \code{x}.
\code{anyNA()} always returns \code{TRUE} or \code{FALSE}, never \code{NA} and
never anything other than a single value.
}
\description{
Valid datey system ranges:
\itemize{
\item Valid dates are from the start of 1000 to the start of 3000.
\item Valid durations are 2000 years or less in magnitude.
}

Values outside the above ranges are treated as NA.

\code{is.na()} tests whether a \code{datey}, \code{durationy} or
\code{datey_interval} is NA by element.

\code{anyNA()} tests whether any element of a \code{datey}, \code{durationy} or
\code{datey_interval} is NA.

For convenience,
\itemize{
\item the constants \link{NA_datey_}, \link{NA_durationy_} and \link{NA_datey_interval_}
are the \code{datey}, \code{durationy} and \code{datey_interval}
versions of NA respectively, and
\item \link[=integer_constants]{integer constants} describing the above valid ranges
are also provided.
}

For performance reasons, intermediate \strong{datey} system
calculations are \emph{not} required to check for NAs.

Throughout the \strong{datey} package, \code{NA} will cause an error when used where
a \code{datey}, \code{durationy} or \code{datey_interval} is expected.
This is because its type is \code{logical} and potentially indicates user
error. If you want an NA value with a \strong{datey} system type, use one of \code{NA_datey_}, \code{NA_durationy_}
or \code{NA_datey_interval_}.
}
\examples{
  t <- c(NA_datey_, datey(2000), datey(999.99, strict = FALSE))
  is.na(t)
  anyNA(t)

  d <- c(NA_durationy_, durationy(1.5))
  is.na(d)
  anyNA(d)

  i <- c(NA_datey_interval_, 2000 \%to\% 2001)
  is.na(i)
  anyNA(i)
}
\seealso{
\link{NA_datey_}, \link{NA_durationy_}, \link{NA_datey_interval_}, \link{integer_constants},
\link{datey}, \link{durationy}, \link{datey_interval}
}
\keyword{NA}
