% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datey.R, R/datey_interval.R, R/durationy.R
\name{is_type}
\alias{is_type}
\alias{is_datey}
\alias{is_datey_interval}
\alias{is_durationy}
\title{Is \code{x} a \code{datey} or a \code{durationy}?}
\usage{
is_datey(x)

is_datey_interval(x)

is_durationy(x)
}
\arguments{
\item{x}{The object to test.}
}
\description{
\code{is_datey()} checks whether an object is a \code{datey}.

\code{is_durationy()} checks whether an object is a \code{durationy}.
}
