% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datey.R
\name{is_xxx_day}
\alias{is_xxx_day}
\alias{is_start_day}
\alias{is_mid_day}
\title{Is a \code{datey} the start (or end) or middle of a day?}
\usage{
is_start_day(x)

is_mid_day(x)
}
\arguments{
\item{x}{The \code{datey} to test.}
}
\value{
A vector of \code{logical}. Elements of the \code{datey} vector that are
\code{NA_datey_} will result in \code{NA} elements of the result vector.
}
\description{
\code{is_start_day()} checks whether \code{x} is the start or end of a day.

\code{is_mid_day()} checks whether \code{x} is the middle of a day.

These properties are \emph{not} necessarily preserved when a duration of
\emph{n} years is added or subtracted.
}
\examples{
# Create (NA, 0 days, 1/4 day, 1/2 day):
t <- datey(c(NA_real_, 2000, 2000 + 0.25/366, 2000 + 0.5/366))
t # <NA>  2000-01-01.0  2000-01-01.25 2000-01-01.5

is_start_day(t) # NA  TRUE FALSE FALSE
is_mid_day(t)   # NA FALSE FALSE  TRUE

# Properties are not necessarily preserved between years:
t <- start_day(2000,7,1) # Leap year
is_start_day(t) # TRUE
is_start_day(t + durationy(1)) # FALSE
}
\seealso{
\link{datey}
}
