% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{mean}
\alias{mean}
\alias{mean.datey}
\alias{mean.durationy}
\title{Mean value of \code{datey} or \code{durationy}}
\usage{
\method{mean}{datey}(x, ..., na.rm = FALSE)

\method{mean}{durationy}(x, ..., na.rm = FALSE)
}
\arguments{
\item{x}{The \code{datey} or \code{durationy}.}

\item{...}{Ignored.}

\item{na.rm}{A logical (\code{TRUE} or \code{FALSE}) indicating whether NA values should be removed before the computation.}
}
\description{
Mean value of \code{datey} or \code{durationy}
}
