% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{mean}
\alias{mean}
\alias{mean.datey}
\alias{mean.durationy}
\title{Mean value of \code{datey} or \code{durationy}}
\usage{
\method{mean}{datey}(x, ..., na.rm = FALSE)

\method{mean}{durationy}(x, ..., na.rm = FALSE)
}
\arguments{
\item{x}{The \code{datey} or \code{durationy}.}

\item{...}{Other arguments (not used in this package).}

\item{na.rm}{A logical (\code{TRUE} or \code{FALSE}) indicating whether NA values should be removed before the computation.}
}
\value{
A scalar of \code{datey} or \code{duration} as appropriate.
}
\description{
Gets the mean value of a vector of \code{datey} or \code{durationy} as a scalar.

This will entail rounding if the mean of
the underlying click counts is not an integer.
}
\examples{
    t <- datey(2000:2003)
    t
    mean(t)
}
