% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/durationy.R
\docType{data}
\name{NA_durationy_}
\alias{NA_durationy_}
\title{The \code{durationy} version of NA}
\format{
An object of class \code{datey_type} (inherits from \code{durationy}) of length 1.
}
\usage{
NA_durationy_
}
\description{
The \code{durationy} version of NA
}
\examples{
  is_durationy(NA_durationy_) # TRUE
}
\seealso{
\link{is_NA}, \link{NA_datey_}, \link{integer_constants}
}
\keyword{datasets}
