% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datey.R, R/datey_interval.R, R/durationy.R
\name{NAs}
\alias{NAs}
\alias{NA_datey_}
\alias{NA_datey_interval_}
\alias{NA_durationy_}
\title{The \code{datey}, \code{durationy} and \code{datey_interval} versions of NA}
\usage{
NA_datey_

NA_datey_interval_

NA_durationy_
}
\value{
Each is a scalar holding the missing-value (\code{NA})
representation of its type: \code{NA_datey_} is a \code{datey}, \code{NA_durationy_} is a
\code{durationy} and \code{NA_datey_interval_} is a \code{datey_interval}.
}
\description{
Throughout the \strong{datey} package, \code{NA} will cause an error when used where
a \code{datey_}, \code{durationy_} or \code{datey_interval_} is expected.
This is because its type is \code{logical} and potentially indicates user
error. If you want an NA value with a \strong{datey} system type, use one of \code{NA_datey_}, \code{NA_durationy_}
or \code{NA_datey_interval_}.
}
\examples{
  is_datey(NA_datey_)
  is.na(NA_datey_)
  is_durationy(NA_durationy_)
  is.na(NA_durationy_)
  is_datey_interval(NA_datey_interval_)
}
\seealso{
\link{is_NA}, \link{integer_constants}, \link{datey}, \link{durationy}, \link{datey_interval}
}
