% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{subset}
\alias{subset}
\alias{[.datey}
\alias{[.durationy}
\alias{[.datey_interval}
\alias{[<-.datey}
\alias{[<-.durationy}
\alias{[<-.datey_interval}
\title{Subset \code{datey}, \code{durationy} or \code{datey_interval} vectors}
\usage{
\method{[}{datey}(x, i, ...)

\method{[}{durationy}(x, i, ...)

\method{[}{datey_interval}(x, i, ...)

\method{[}{datey}(x, i) <- value

\method{[}{durationy}(x, i) <- value

\method{[}{datey_interval}(x, i) <- value
}
\arguments{
\item{x}{A \code{datey}, \code{durationy} or \code{datey_interval}.}

\item{i}{Indices to extract.}

\item{...}{Other arguments.}

\item{value}{Value to assign.}
}
\value{
The subset.
}
\description{
Subsets \code{datey}, \code{durationy} or \code{datey_interval} vectors.
}
\examples{
  x <- datey(2001:2004)
  x
  x[2:3]
  x[2:3] <- datey(1999)
  x
}
\seealso{
\link{datey}, \link{durationy}, \link{datey_interval}, \link{combine}
}
\keyword{subset}
