% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datey.R
\name{text_from_datey}
\alias{text_from_datey}
\alias{as.character.datey}
\alias{format.datey}
\alias{print.datey}
\title{Format or print a \code{datey}}
\usage{
\method{as.character}{datey}(x, ...)

\method{format}{datey}(x, include_day_fraction = TRUE, ...)

\method{print}{datey}(x, include_day_fraction = TRUE, max = NULL, ...)
}
\arguments{
\item{x}{The \code{datey} to print or format.}

\item{...}{Other arguments.}

\item{include_day_fraction}{Whether to include the fractional day part.
Defaults to \code{FALSE}.}

\item{max}{Numeric or \code{NULL}, specifying the maximal number of entries to be
printed. When \code{NULL}, \code{getOption("max.print")} used. Defaults to \code{NULL}.}
}
\description{
A \code{datey} is printed as either
\itemize{
\item \code{YYYY-MM-DD}, i.e. ISO 8601 extended date format, or
\item \code{YYYY-MM-DD.FFF} where \code{.FFF} is the day fraction part
}

If \code{include_day_fraction} is \code{TRUE} then \code{.FFF} is included even if it is 0.

Note that a \code{datey} created as the end of a day (or with day fraction 1) will
print as the start of the following day.
}
