% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datey.R
\name{xxx_day}
\alias{xxx_day}
\alias{start_day}
\alias{mid_day}
\alias{end_day}
\alias{as_start_day}
\alias{as_mid_day}
\alias{as_end_day}
\title{Create a \code{datey} for the start, middle or end of a day}
\usage{
start_day(year, month, day, strict = TRUE)

mid_day(year, month, day, strict = TRUE)

end_day(year, month, day, strict = TRUE)

as_start_day(x, ...)

as_mid_day(x, ...)

as_end_day(x, ...)
}
\arguments{
\item{year}{Calendar year.
Valid years are from 1000 (inclusive) to 3000 (exclusive).
If provided as \code{double} then these \emph{must be integers}.}

\item{month}{Month number in calendar year, with 1 representing January.
If provided as \code{double} then these \emph{must be integers}.}

\item{day}{Day number in month, with 1 representing the first day of the month.
If provided as \code{double} then these \emph{must be integers}.}

\item{strict}{How to handle calendar years less than 1000 or greater than or equal to 3000 and day
fractions not in the interval [0,1].
\itemize{
\item If \code{strict} is \code{TRUE} -- the default -- then execution is stopped.
\item If \code{strict} is \code{FALSE} then \code{NA} is returned.
}

(NA arguments result in NA regardless of \code{strict}.)}

\item{day_fraction}{The fraction of the day, in [0,1].
0 means the start of the day,
0.5 means the middle of the day, and
1 means the end of the day
(which is identical to the start of the next day).}
}
\description{
Create a \code{datey} for the start, middle or end of the day
specified by calendar year, month and day.

The lengths of vector arguments must be multiples of each other.

To deconstruct a \code{datey}, use \code{\link[=to_ymdf]{to_ymdf()}}.
}
\section{Edge cases}{
The following are special cases:
\enumerate{
\item \code{end_day(999, 12, 31)} \emph{will} cause an error,
even though in theory it represents the legal \code{datey} 1000-01-01.0.
\item \code{end_day(2999, 12, 31)} will \emph{not} cause an error,
even though in theory it represents the illegal \code{datey}
3000-01-01.0. The resulting \code{datey} will be invalid though.
}
}

