% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datey.R
\name{xxx_day}
\alias{xxx_day}
\alias{start_day}
\alias{mid_day}
\alias{end_day}
\title{Create a \code{datey} for the start, middle or end of a day}
\usage{
start_day(year, month, day, strict = TRUE)

mid_day(year, month, day, strict = TRUE)

end_day(year, month, day, strict = TRUE)
}
\arguments{
\item{year}{Calendar year.
Valid years are from 1000 to 3000 (although the only legal date in 3000
is the start of 3000-01-01).}

\item{month}{Month number in calendar year, with 1 representing January.
If provided as \code{double} then these \emph{must be integers}.}

\item{day}{Day number in month, with 1 representing the first day of the month.
If provided as \code{double} then these \emph{must be integers}.}

\item{strict}{How to handle calendar years less than 1000 or greater than 3000 and day
fractions not in the interval [0,1].
\itemize{
\item If \code{strict} is \code{TRUE} -- the default -- then execution is stopped.
\item If \code{strict} is \code{FALSE} then \code{NA} is returned.
}

NA arguments result in NA (and do not stop execution) regardless of \code{strict}.}
}
\description{
Create a \code{datey} for the start, middle or end of the day
specified by calendar year, month and day.

The lengths of vector arguments must be multiples of each other.
}
\examples{
  start_day(1999, 12, 31)
  mid_day(1999, 12, 31)
  end_day(1999, 12, 31)
}
\seealso{
Use \code{\link[=as_start_day]{as_start_day()}}, \code{\link[=as_mid_day]{as_mid_day()}} or \code{\link[=as_end_day]{as_end_day()}} to to create a
\code{datey} from a base R date or datetime.

Use \code{\link[=datey]{datey()}} to to create a \code{datey} direct from fractional calendar years.

To deconstruct a \code{datey}, use \code{\link[=to_ymdf]{to_ymdf()}}.
}
