% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datey.R
\name{ymdf}
\alias{ymdf}
\alias{to_ymdf}
\alias{from_ymdf}
\title{Create or decompose a \code{datey} using calendar year, month, day and day fraction}
\usage{
to_ymdf(datey)

from_ymdf(year, month, day, day_fraction, strict = TRUE)
}
\arguments{
\item{datey}{A \code{datey} to be deconstructed.}

\item{year}{Calendar year.
Valid years are from 1000 to 3000 (although the only legal date in 3000
is the start of 3000-01-01).
If provided as \code{double} then these \emph{must be integers}.}

\item{month}{Month number in calendar year, with 1 representing January.
If provided as \code{double} then these \emph{must be integers}.}

\item{day}{Day number in month, with 1 representing the first day of the month.
If provided as \code{double} then these \emph{must be integers}.}

\item{day_fraction}{The fraction of the day, in [0,1].
0 means the start of the day,
0.5 means the middle of the day, and
1 means the end of the day
(which is identical to the start of the next day).}

\item{strict}{How to handle invalid arguments.
\itemize{
\item If \code{strict} is \code{TRUE} -- the default -- then execution is stopped.
\item If \code{strict} is \code{FALSE} then \code{NA} is returned.
}

NA arguments result in NA (and do not stop execution) regardless of \code{strict}.}
}
\description{
The lengths of vector arguments must be multiples of each other.

\code{to_ymdf()} returns a list of the \code{year}, \code{month}, \code{day} and \code{day_fraction}
breakdown of a \code{datey}, where
\itemize{
\item \code{year} is an \code{integer} in [1000,3000],
\item \code{month} is an \code{integer} in [1,12],
\item \code{day} is an \code{integer} in [1,N], where N is the number of days
in the month specified by \code{year} and \code{month}, and
\item \code{day_fraction} is a \code{double} in [0,1) representing the fraction of the day,
where e.g. 0 means the start and 0.5 means the middle of the day.
}

Alternatively, if you want only one or two components, use the list-like
syntax \link[=datey_properties]{\verb{$year}, \verb{$month}, \verb{$day} or \verb{$day_fraction}}.

If the \code{datey} was constructed using \code{end_day} or \code{day_fraction = 1} then
\code{to_ymdf()} will return the \emph{start} of the \emph{next} day with \code{day_fraction = 0}.

\code{from_ymdf()} creates a \code{datey} from a calendar year, month, and day
fraction. In practice, prefer one of \code{\link[=start_day]{start_day()}}, \code{\link[=mid_day]{mid_day()}} or
\code{\link[=end_day]{end_day()}} for clarity.
}
\examples{
  t <- from_ymdf(2001, 2, 3, 0.5)
  t
  to_ymdf(t)
  t$year
  t$month
  t$day
  t$day_fraction
}
\seealso{
Use \code{\link[=datey]{datey()}} to create a \code{datey} direct from years or a base R date.

Use the syntax \href{datey_properties}{\verb{$year}, \verb{$month}, \verb{$day} or \verb{$day_fraction}}
to extract one component at a time.
}
